Refactor: [User, Post] API 공통응답 및 validation 적용

+++ services/postService.js
@@ -5,6 +5,14 @@ const Post = require('../models/Post');
 // 게시글 생성
 const createPost = async ({ userId, title, content, tags }) => {
     const db = getDb();
+
+    // 사용자 존재 여부 확인
+    const user = await db.collection('users').findOne({ _id: new ObjectId(userId) });
+    if (!user) {
+        throw new Error('존재하지 않는 사용자입니다.');
+    }
+
+    // todo: 태크가 존재하는지 (나중에 class 추가하든지 해야할듯)
     const newPost = new Post(userId, title, content, tags);
     const result = await db.collection('posts').insertOne(newPost);
     return { _id: result.insertedId, ...newPost };
@@ -28,8 +36,23 @@ const getPostById = async (postId) => {
 };
 
 // 게시글 수정
-const updatePost = async (postId, { title, content, tags }) => {
+const updatePost = async (postId, { userId, title, content, tags }) => {
     const db = getDb();
+
+    const user = await db.collection('users').findOne({ _id: new ObjectId(userId) });
+    if (!user) {
+        throw new Error('존재하지 않는 사용자입니다.');
+    }
+
+    const post = await db.collection('posts').findOne({ _id: new ObjectId(postId) });
+    if (!post) {
+        throw new Error('존재하지 않는 게시글입니다.');
+    }
+
+    if (post.userId !== userId) {
+        throw new Error('해당 게시글의 작성자가 아닙니다.');
+    }
+
     const updatedPost = {
         ...(title && { title }),
         ...(content && { content }),
@@ -43,22 +66,32 @@ const updatePost = async (postId, { title, content, tags }) => {
     );
 
     if (result.matchedCount === 0) {
-        throw new Error(`Post ${postId} not found`);
+        throw new Error('존재하지 않는 게시글입니다.');
     }
 
-    return { message: `Post ${postId} updated`, updatedPost };
+    return updatedPost;
 };
 
 // 게시글 삭제
-const deletePost = async (postId) => {
+const deletePost = async (postId, userId) => {
     const db = getDb();
+
+    const user = await db.collection('users').findOne({ _id: new ObjectId(userId) });
+    if (!user) {
+        throw new Error('존재하지 않는 사용자입니다.');
+    }
+
+    if (post.userId !== userId) {
+        throw new Error('해당 게시글의 작성자가 아닙니다.');
+    }
+
     const result = await db.collection('posts').deleteOne({ _id: new ObjectId(postId) });
 
     if (result.deletedCount === 0) {
-        throw new Error(`Post ${postId} not found`);
+        throw new Error("존재하지 않는 게시글입니다.");
     }
 
-    return { message: `Post ${postId} deleted` };
+    return "게시글이 성공적으로 삭제되었습니다.";
 };
 
 // // 댓글 추가 (내장 문서 사용)
